refactor: extract shared crypto/key helpers
Fixes #19
