+++ src/keys.ts
@@ -1,0 +1,8 @@
+export function generateProxyKey(): string {
+  const randomBytes = crypto.getRandomValues(new Uint8Array(24));
+  const base64 = btoa(String.fromCharCode(...randomBytes))
+    .replace(/\+/g, "-")
+    .replace(/\//g, "_")
+    .replace(/=/g, "");
+  return "cpk_" + base64;
+}
